fix(lint): split accessor-accessibility helpers to sibling module under size ratchet
